[2.0] Add Laravel\Dusk\Concerns\ProvidesBrowser
Extract reusable code from Laravel\Dusk\TestCase to make it easily for
user to either build standalone browser tests suite to test local or
external web application. For my case, I'm using similar class to allow
me to run application tests on production environment right after doing
server/database maintenance.
